backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-community
langchain-google-genai
langgraph
faiss-cpu
sqlite-utils
pytest
httpx
cryptography
watchfiles
jinja2
python-multipart
starlette
typing_extensions
google-generativeai
